+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.3.0-0.22b0...HEAD)
+- Include Flask 2.0 as compatible with existing flask instrumentation
+  ([#545](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/545))
 
 ### Changed
 - `opentelemetry-instrumentation-tornado` properly instrument work done in tornado on_finish method.
@@ -17,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#469](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/469))
 - Implemented specification that HTTP span attributes must not contain username and password.
   ([#538](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/538))
+- Changed the psycopg2-binary to psycopg2 as dependency in production
+  ([#543](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/543))
 
 ### Added
 - `opentelemetry-instrumentation-httpx` Add `httpx` instrumentation